popbl_servicesapp/flask_app/auth/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.9.2
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.6
Werkzeug==0.15.6
pika==1.1.0
bcrypt==3.1.7
pyjwt==1.7.1
cryptography==2.8
pyopenssl==19.0.0
# Consul library
consulate==0.6.0
flask-consulate==0.2.0
# To get my IP
netifaces==0.10.9